Move files and change path specifier - path specifier: relative -> non-relative - Fixed #154
