+++ src/libs/systems/PackageJsonReplacer/index.ts
@@ -1,4 +1,4 @@
-import { ProjectOption } from '../../models/ProjectOptions';
+import { ProjectOption } from 'src/models/ProjectOptions';
 
 type PackageJson = {
   name: string;
