+++ src/models/Result/index.ts
